requirements.txt
pymupdf>=1.24.1
pdfplumber>=0.11.0
pillow>=10.3.0
openai>=1.30.1
groq>=0.5.0

python-dotenv>=1.0.1
jinja2>=3.1.4
streamlit>=1.35.0
pytesseract>=0.3.10
requests>=2.32.2
weasyprint>=60.0 ; sys_platform != "win32"

